Update README.md
Reference:
Hamid Mohammadi, Vahid Morovati, Emad Poshtan, Roozbeh Dargazany
Understanding decay functions and their contribution in modeling of thermal-induced aging of cross-linked polymers
Polymer Degradation and Stability
https://doi.org/10.1016/j.polymdegradstab.2020.109108

Summary:
The model describes the aging induced damage with respect to experimental studies on the process of chemical aging which suggests decomposition of original network and creation of a new one. Accordingly, in the course of aging, the strain energy of the polymer matrix is divided from two independent sources, (i) a decomposing original matrix (ii) a newly formed matrix. The model is validated with respect to a comprehensive set of own experimental data as well as data available in literature that were designed to capture thermal induced aging effects of constitutive behavior of polymers.